# docs/openapi-v3-api-docs.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "ProjectKu Web API",
    "description" : "REST API 文档",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/api"
  } ],
  "paths" : {
...
    "/v1/reviews" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "productId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "orderId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "review-controller" ],
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/register" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ReviewCreateRequest" : {
        "type" : "object",
        "properties" : {
          "orderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "productId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "rating" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "string",
            "example" : "很好用，物流很快"
          },
          "images" : {
            "type" : "string",
            "example" : "[]"
          }
        }
      },
...
      "RegisterRequest" : {
        "type" : "object",
        "properties" : {
          "account" : {
            "type" : "string",
            "example" : "17863026867@163.com"
          },
          "password" : {
            "type" : "string",
            "example" : "bill0821"
          },
          "nickname" : {
            "type" : "string",
            "example" : "Bill"
          }
        }
      },
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "account" : {
            "type" : "string",
            "example" : "17863026867@163.com"
          },
          "password" : {
            "type" : "string",
            "example" : "bill0821"
          }
        }
      },
...
    }
  }
}